requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.3
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.0.3
cssutils==1.0.2
dj-database-url==0.5.0
Django==2.1.5
django-appconf==1.0.2
django-cors-middleware==1.3.1
django-extensions==2.1.4
django-heroku==0.3.1
django-inlinecss==0.1.2
django-rest-swagger==2.1.0
django-staticfiles==1.2.1
djangorestframework==3.9.0
flake8==3.6.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mccabe==0.6.1
mock==2.0.0
more-itertools==5.0.0
openapi-codec==1.3.2
pbr==5.1.1
pluggy==0.8.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
py==1.7.0
pycodestyle==2.4.0
pyflakes==2.0.0
PyJWT==1.7.1
pynliner3==0.6
pytest==4.1.0
python-coveralls==2.9.1
pytz==2018.9
PyYAML==3.13
requests==2.21.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.7.2
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
